Swapping two tasks now works.
A visual effect also occurs to show the user that a task has been
selected for swapping.

removeTaskEvent has been changed to rebuild the GUI after a task has
been removed in order to prevent bugs involving task IDs not being the
same as their index in taskListBox's children collection.